pyproject.toml
[tool.poetry]
name = "auto_eia"
version = "0.1.0"
description = ""
authors = ["Max Dumas <max@dumas.nyc>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "2.50.0"
"aws-cdk.aws-apigatewayv2-alpha" = "2.50.0-alpha.0"
"aws-cdk.aws-apigatewayv2-integrations-alpha" = "2.50.0-alpha.0"
"aws-cdk.aws-lambda-python-alpha" = "2.50.0-alpha.0"
constructs = "^10"
boto3 = "^1.26.4"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
black = "^22.10.0"
mypy = "^0.990"
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
